+++ src/Attributes/SequenceField.php
@@ -30,7 +30,9 @@ class SequenceField implements TypeField, SupportsScopes
 
     public function acceptsType(string $type): bool
     {
-        return $type === 'array';
+        return $type === 'array'
+            || $type === 'iterable'
+            || is_a($type, \Traversable::class, true);
     }
 
     public function shouldImplode(): bool