+++ CHANGELOG.md
@@ -4,7 +4,27 @@ All notable changes to `Serde` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 0.5.0 - YYYY-MM-DD
+## 0.6.0 - YYYY-MM-DD
+
+### Added
+- Support for iterable properties, including generators.
+- Support for serializing/deserializing from CSV files.
+- Support for stream-serializing to a CSV format.
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
+
+## 0.5.0 - 2022-07-22
 
 ### Added
 - Dictionary fields can now be restricted to just string or just integer keys.